properly configure faye types
