+++ tsconfig.json
@@ -20,10 +20,13 @@
       "@/*": [
         "./src/*"
       ]
-    }
+    },
+    "typeRoots": [
+      "./node_modules/@types",
+      "./src/@types"
+    ]
   },
   "include": [
-    "@types",
     "src"
   ],
   "exclude": [
